requirements/base.txt
asgi-correlation-id==4.3.4
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.0
dependency-injector==4.46.0
pydantic_settings==2.9.1
uuid7
SQLAlchemy
psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.16.1
pyjwt[crypto]==2.10.1
cryptography==45.0.4
pytest 
pytest-asyncio
pytest-cov
